feat: implement functionality to use status isLoading and sendRequest from Redux state.
