+++ react-redux/src/compoments/articles-holder/ArticlesHolder.module.scss
@@ -3,32 +3,3 @@
 .article {
   @include flex();
 }
-
-.pagination {
-  @include flex($direction: row);
-  max-width: 250px;
-  margin: 10px 0;
-
-  > * {
-    margin: 0 5px;
-  }
-}
-
-.pagination button {
-  border-radius: 5px;
-  background-color: $yellow;
-  font-weight: 800;
-  color: $dark;
-
-  &:active {
-    background-color: $pink;
-  }
-}
-
-.pagination input {
-  @include font($size: 2rem, $weight: 800);
-  border-radius: 5px;
-  background-color: $dark;
-  color: $yellow;
-  border: 1px solid $yellow;
-}
