+++ react-redux/src/compoments/articles-holder/ArticlesHolder.module.scss.d.ts
@@ -2,7 +2,6 @@
 // Please do not change this file!
 interface CssExports {
   'article': string;
-  'pagination': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
